Calculate scroll position from the bottom
Previously, when scrolling to the top to fetch messages, the new scroll
position would be calculated based on the number of messages and their
heights.

With this commit, the new scroll position is based on the distance from
the bottom before messages are prepended, since that value should be
equal before and after.

E.g. if the scroll position is 1000px from the bottom before the new
messages are prepended, we make sure the new scroll position is also
1000px from the bottom after they're prepended.