Calculate speech height based on text width
Uses canvas and its measureText method to find the length of text in
each message, divide that by the length of the container, and round up
to get the number of lines for each message. Then use that to define a
fixed height for all children, and pass that list of heights to
react-infinite.